refactor: reorganize project structure by replacing utils.py with dedicated modules CORE_Cache, CORE_GTKUtils, CORE_Helpers, and CORE_Networking.
