+++ src/CORE_Networking.py
@@ -1,0 +1,54 @@
+import threading
+from threading import Lock
+from gi.repository import GLib
+from .CORE_Helpers import get_time_difference
+
+_fetch_callbacks = []
+_fetch_lock = Lock()
+
+def fetch_all_weather_data_async(on_success=None, on_error=None):
+    """
+    Fetches all weather data in a thread-safe manner.
+    """
+    from .CORE_weatherData import (
+        fetch_current_weather,
+        fetch_hourly_forecast,
+        fetch_daily_forecast,
+        fetch_current_air_pollution
+    )
+
+    with _fetch_lock:
+        if on_success:
+            _fetch_callbacks.append(on_success)
+        if any(t.name == "fetch_all_worker" for t in threading.enumerate()):
+            return
+
+    def _worker():
+        try:
+            cwd = threading.Thread(target=fetch_current_weather, name="cwt")
+            cwd.start()
+            cwd.join()
+
+            threads = [
+                threading.Thread(target=fetch_hourly_forecast, name="hft"),
+                threading.Thread(target=fetch_daily_forecast, name="dft"),
+                threading.Thread(target=fetch_current_air_pollution, name="apt"),
+                threading.Thread(target=get_time_difference, args=("", True), name="local_time")
+            ]
+            
+            for t in threads: t.start()
+            for t in threads: t.join()
+            
+            with _fetch_lock:
+                callbacks = _fetch_callbacks[:]
+                _fetch_callbacks.clear()
+            
+            for cb in callbacks:
+                GLib.idle_add(cb)
+                
+        except Exception as e:
+            print(f"Error fetching data: {e}")
+            if on_error:
+                GLib.idle_add(on_error, str(e))
+
+    threading.Thread(target=_worker, name="fetch_all_worker", daemon=True).start()
