+++ src/UI_CompDrawLineGraph.py
@@ -15,7 +15,7 @@ class LineGraph:
         self.dw = Gtk.DrawingArea()
         self.dw.set_size_request(260, 120) # Increased height for date labels
         
-        from .utils import safe_set_draw_func
+        from .CORE_GTKUtils import safe_set_draw_func
         safe_set_draw_func(self.dw, self, "draw")
 
     def draw(self, area, ctx, width, height, data):
