+++ src/windowLocations.py
@@ -7,7 +7,7 @@ gi.require_version("Gtk", "4.0")
 gi.require_version("Adw", "1")
 from gi.repository import Gtk, Adw
 
-from .utils import create_toast
+from .CORE_Helpers import create_toast
 from .API_FindCity import find_city
 from .config import settings
 from gettext import gettext as _, pgettext as C_
